make tmt.py a executable script and add color flag
