+++ tmt.py
@@ -1,6 +1,10 @@
+#!/usr/bin/env python3
 import argparse
+import os
+import sys
 import pathlib
 
+from internal import __version__
 from internal.commands.verify import command_verify_config, command_verify_verdicts
 from internal.context import TMTContext, find_problem_dir
 from internal.commands import (
@@ -12,24 +16,32 @@ from internal.commands import (
     command_verify,
 )
 from internal.exceptions import TMTMissingFileError, TMTInvalidConfigError
-from internal import __version__
-from internal.formatting import TerminalFormatter
+from internal.formatting import TerminalFormatter, PlainFormatter
 from internal.verify.verifier import TMTVerifyIssueType
 
 
 def main():
-    parser = argparse.ArgumentParser(description="TMT - Task Management Tools")
+    shared = argparse.ArgumentParser(add_help=False)
+    shared.add_argument("--color", choices=["always", "auto", "never"], default="auto")
+
+    parser = argparse.ArgumentParser(
+        description="TMT - Task Management Tools", parents=[shared]
+    )
     parser.add_argument(
         "--version",
         action="version",
         version=f"TMT {__version__}",
         help="Show the version of TMT.",
     )
+
     subparsers = parser.add_subparsers(dest="command", required=True)
 
     # parser_init = subparsers.add_parser("init", help="Init a TMT problem directory.")
 
-    parser_gen = subparsers.add_parser("gen", help="Generate testcases.")
+    # tmt gen
+    parser_gen = subparsers.add_parser(
+        "gen", help="Generate testcases.", parents=[shared]
+    )
     parser_gen.add_argument(
         "-r",
         "--show-reason",
@@ -42,42 +54,75 @@ def main():
         help="Check if the hash digest of the testcases matches.",
     )
 
-    parser_invoke = subparsers.add_parser("invoke", help="Invoke a solution.")
+    # tmt invoke
+    parser_invoke = subparsers.add_parser(
+        "invoke", help="Invoke a solution.", parents=[shared]
+    )
     parser_invoke.add_argument("-r", "--show-reason", action="store_true")
     parser_invoke.add_argument("submission_files", nargs="*")
 
+    # tmt clean
     parser_clean = subparsers.add_parser(
-        "clean", help="Clean-up a TMT problem directory."
+        "clean", help="Clean-up a TMT problem directory.", parents=[shared]
     )
     parser_clean.add_argument(
         "-y", "--yes", action="store_true", help="Automatic yes to prompts."
     )
 
-    parser_export = subparsers.add_parser("export", help="Export packages")
+    # tmt export
+    parser_export = subparsers.add_parser(
+        "export", help="Export packages", parents=[shared]
+    )
     parser_export.add_argument("output", help="The filename of the exported zip file.")
 
-    subparsers.add_parser("make-public", help="Build public attachment archive file.")
+    # tmt make-public
+    subparsers.add_parser(
+        "make-public", help="Build public attachment archive file.", parents=[shared]
+    )
 
-    parser_verify = subparsers.add_parser("verify", help="Check issues.")
+    # tmt verify ...
+    parser_verify = subparsers.add_parser(
+        "verify", help="Check issues.", parents=[shared]
+    )
     verify_subparser = parser_verify.add_subparsers(
         dest="issue_class", help="The issue class to be verified."
     )
-    verify_subparser.add_parser("all", help="Verify all issue classes.")
+    # tmt verify all
+    verify_subparser.add_parser(
+        "all", help="Verify all issue classes.", parents=[shared]
+    )
+    # tmt verify verdicts
     parser_verify_verdicts = verify_subparser.add_parser(
-        "verdicts", help="Verify solution verdicts."
+        "verdicts", help="Verify solution verdicts.", parents=[shared]
     )
     parser_verify_verdicts.add_argument(
         "-s", "--solution", help="The solution filename in solutions/."
     )
-    verify_subparser.add_parser("config", help="Verify configs.")
+    # tmt verify config
+    verify_subparser.add_parser("config", help="Verify configs.", parents=[shared])
 
     args = parser.parse_args()
+
+    # forced by args
+    if args.color == "always":
+        formatter = TerminalFormatter()
+    elif args.color == "never":
+        formatter = PlainFormatter()
+    # environment variable
+    elif os.getenv("NO_COLOR") or os.getenv("TERM") == "dumb":
+        formatter = PlainFormatter()
+    elif os.getenv("FORCE_COLOR"):
+        formatter = TerminalFormatter()
+    # fallback to terminal detection
+    elif os.isatty(sys.stdout):
+        formatter = TerminalFormatter()
+    else:
+        formatter = PlainFormatter()
 
     if args.command == "init":
         print("Directory initialization is not implemented yet.")
         return
 
-    formatter = TerminalFormatter()
     cwd = pathlib.Path.cwd()
     problem_dir = find_problem_dir(cwd)  # TODO specify it in args
     script_dir = str(pathlib.Path(__file__).parent.resolve())
